python2 support

+++ GC_analysis/GC_analysis.py
@@ -173,12 +173,12 @@ def generate_result():
     for i in range((seq_len - window_size + shift) // shift):  # Iterate over the total number of shifts
         frag = record.seq[i * shift: i * shift + window_size]  # Extract the string for counting
         # Count number of C and G and convert to percentage
-        percent = round((frag.count("C") + frag.count("G")) / window_size * 100)
+        percent = int(round((frag.count("C") + frag.count("G")) / float(window_size) * 100))
         write_content(i * shift, percent)
     if (i + 1) * shift < seq_len and not omit_tail:
         # if trailing sequence exits and omit_tail is False
         frag = record.seq[(i + 1) * shift:]
-        percent = round((frag.count("C") + frag.count("G")) / len(frag) * 100)
+        percent = int(round((frag.count("C") + frag.count("G")) / float(len(frag)) * 100))
         write_content((i + 1) * shift, percent)
